♻️ refactor: componentize duplicate codes on registration and login screens
