+++ client/src/components/Auth/Registration.tsx
@@ -3,14 +3,13 @@ import { useNavigate } from 'react-router-dom';
 import React, { useState, useEffect } from 'react';
 import { useRegisterUserMutation, useGetStartupConfig } from 'librechat-data-provider/react-query';
 import type { TRegisterUser } from 'librechat-data-provider';
-import { GoogleIcon, FacebookIcon, OpenIDIcon, GithubIcon, DiscordIcon } from '~/components';
-import { ThemeSelector } from '~/components/ui';
-import SocialButton from './SocialButton';
 import { useLocalize } from '~/hooks';
+import { ErrorMessage } from './ErrorMessage';
+import AuthLayout from './AuthLayout';
 
 const Registration: React.FC = () => {
   const navigate = useNavigate();
-  const { data: startupConfig } = useGetStartupConfig();
+  const { data: startupConfig, error: startupConfigError, isFetching } = useGetStartupConfig();
   const localize = useLocalize();
 
   const {
@@ -45,12 +44,6 @@ const Registration: React.FC = () => {
     }
   }, [startupConfig, navigate]);
 
-  if (!startupConfig) {
-    return null;
-  }
-
-  const socialLogins = startupConfig.socialLogins ?? [];
-
   const renderInput = (id: string, label: string, type: string, validation: object) => (
     <div className="mb-2">
       <div className="relative">
@@ -83,120 +76,21 @@ const Registration: React.FC = () => {
     </div>
   );
 
-  const providerComponents = {
-    discord: (
-      <SocialButton
-        key="discord"
-        enabled={startupConfig.discordLoginEnabled}
-        serverDomain={startupConfig.serverDomain}
-        oauthPath="discord"
-        Icon={DiscordIcon}
-        label={localize('com_auth_discord_login')}
-        id="discord"
-      />
-    ),
-    facebook: (
-      <SocialButton
-        key="facebook"
-        enabled={startupConfig.facebookLoginEnabled}
-        serverDomain={startupConfig.serverDomain}
-        oauthPath="facebook"
-        Icon={FacebookIcon}
-        label={localize('com_auth_facebook_login')}
-        id="facebook"
-      />
-    ),
-    github: (
-      <SocialButton
-        key="github"
-        enabled={startupConfig.githubLoginEnabled}
-        serverDomain={startupConfig.serverDomain}
-        oauthPath="github"
-        Icon={GithubIcon}
-        label={localize('com_auth_github_login')}
-        id="github"
-      />
-    ),
-    google: (
-      <SocialButton
-        key="google"
-        enabled={startupConfig.googleLoginEnabled}
-        serverDomain={startupConfig.serverDomain}
-        oauthPath="google"
-        Icon={GoogleIcon}
-        label={localize('com_auth_google_login')}
-        id="google"
-      />
-    ),
-    openid: (
-      <SocialButton
-        key="openid"
-        enabled={startupConfig.openidLoginEnabled}
-        serverDomain={startupConfig.serverDomain}
-        oauthPath="openid"
-        Icon={() =>
-          startupConfig.openidImageUrl ? (
-            <img src={startupConfig.openidImageUrl} alt="OpenID Logo" className="h-5 w-5" />
-          ) : (
-            <OpenIDIcon />
-          )
-        }
-        label={startupConfig.openidLabel}
-        id="openid"
-      />
-    ),
-  };
-
-  const privacyPolicy = startupConfig.interface?.privacyPolicy;
-  const termsOfService = startupConfig.interface?.termsOfService;
-
-  const privacyPolicyRender = privacyPolicy?.externalUrl && (
-    <a
-      className="text-sm text-green-500"
-      href={privacyPolicy.externalUrl}
-      target={privacyPolicy.openNewTab ? '_blank' : undefined}
-      rel="noreferrer"
-    >
-      {localize('com_ui_privacy_policy')}
-    </a>
-  );
-
-  const termsOfServiceRender = termsOfService?.externalUrl && (
-    <a
-      className="text-sm text-green-500"
-      href={termsOfService.externalUrl}
-      target={termsOfService.openNewTab ? '_blank' : undefined}
-      rel="noreferrer"
-    >
-      {localize('com_ui_terms_of_service')}
-    </a>
-  );
-
   return (
-    <div className="relative flex min-h-screen flex-col bg-white dark:bg-gray-900">
-      <div className="mt-12 h-24 w-full bg-cover">
-        <img src="/assets/logo.svg" className="h-full w-full object-contain" alt="Logo" />
-      </div>
-      <div className="absolute bottom-0 left-0 md:m-4">
-        <ThemeSelector />
-      </div>
-      <div className="flex flex-grow items-center justify-center">
-        <div className="w-authPageWidth overflow-hidden bg-white px-6 py-4 dark:bg-gray-900 sm:max-w-md sm:rounded-lg">
-          <h1
-            className="mb-4 text-center text-3xl font-semibold text-black dark:text-white"
-            style={{ userSelect: 'none' }}
-          >
-            {localize('com_auth_create_account')}
-          </h1>
-          {error && (
-            <div
-              className="rounded-md border border-red-500 bg-red-500/10 px-3 py-2 text-sm text-gray-600 dark:text-gray-200"
-              role="alert"
-              data-testid="registration-error"
-            >
-              {localize('com_auth_error_create')} {errorMessage}
-            </div>
-          )}
+    <AuthLayout
+      header={localize('com_auth_create_account')}
+      isFetching={isFetching}
+      startupConfig={startupConfig}
+      startupConfigError={startupConfigError}
+    >
+      {error && (
+        <ErrorMessage>
+          {localize('com_auth_error_create')} {errorMessage}
+        </ErrorMessage>
+      )}
+
+      {!startupConfigError && !isFetching && (
+        <>
           <form
             className="mt-6"
             aria-label="Registration form"
@@ -264,39 +158,16 @@ const Registration: React.FC = () => {
               </button>
             </div>
           </form>
+
           <p className="my-4 text-center text-sm font-light text-gray-700 dark:text-white">
             {localize('com_auth_already_have_account')}{' '}
             <a href="/login" aria-label="Login" className="p-1 text-green-500">
               {localize('com_auth_login')}
             </a>
           </p>
-          {startupConfig.socialLoginEnabled && (
-            <>
-              {startupConfig.emailLoginEnabled && (
-                <>
-                  <div className="relative mt-6 flex w-full items-center justify-center border border-t border-gray-300 uppercase dark:border-gray-600">
-                    <div className="absolute bg-white px-3 text-xs text-black dark:bg-gray-900 dark:text-white">
-                      Or
-                    </div>
-                  </div>
-                  <div className="mt-8" />
-                </>
-              )}
-              <div className="mt-2">
-                {socialLogins.map((provider) => providerComponents[provider] || null)}
-              </div>
-            </>
-          )}
-        </div>
-      </div>
-      <div className="align-end m-4 flex justify-center gap-2">
-        {privacyPolicyRender}
-        {privacyPolicyRender && termsOfServiceRender && (
-          <div className="border-r-[1px] border-gray-300 dark:border-gray-600" />
-        )}
-        {termsOfServiceRender}
-      </div>
-    </div>
+        </>
+      )}
+    </AuthLayout>
   );
 };
